Display diagnostic info in console
In order to better understand what version of the source code we are
browsing in a given environment it's important that we know what version
number of the current assembly has been deployed. This helps in the case
where our deploy systems are telling us one thing but the website another.
Such is the case when browsing a cached version of the site. To test this
bring up the console inside a web browser on any page and note the version
number is displayed. It is worth testing the full site to make sure the
low level reflection calls work on pages which behave in unexpected ways.