Remove race condition from postcode file check
Because the `PostcodeFile` property of the `AllowedPostcodeFeature` class
creates a `StreamReader` on each `get`, there is potential for the first
invocation to succeed and the subsequent invocation to fail (if the file
is locked or deleted for example). So by pulling out into an inline
variable it's only invoked once and consumed by the next two lines as
expected.