This adds jumping to the first error on personal details page.
On some browsers it will scroll to the first error, on others
it will jump there. That is expected behaviour, completely fine.
Likewise on some browsers it will position the first error in the middle
of the page, on others it will be at the top. This is fine too.