Fixes me/someone else validation issue
This was quite a rabbit hole to fix. It turns out
that `.valid()` doesn't just return true or false
for whether the field is valid. But it runs the entire
validator including side effects that cause the errorList
to be overwritten.

When I added `:not(.validate-hidden)` to the ignore,
so that DoB could be validated despite being hidden,
that caused the DoB validation to work "properly"
including the side effect.

I have replaced `.valid()` with `isValid(id)` where you
pass over the id of the field and it checks the errorList.
This assumes validation has already taken place.

Given that this triggers either on full form validation or
after each date field has been force validated, it seems to work.