Moving dependencies from files to npm modules
I had an issue where by pulling from npm instead of files, it meant it would run through the entire dependency graph. As we manage it more manually by putting jQuery at the top of the page, this became an issue.

So in the end I've gone with solution of code splitting where pretty much just jQuery goes into common.js with bundle-head and bundle using that. There's a small hack where cookie is now in head so that the code splitting works as expected.

Note that this is due to moving licensed files away so that we can
publicly mirror on GitHub.