This fixes the spinner that didn't spin.
The reason it was broken was that there were new CSS required from the
version in npm of the spinner. It was not easy to add these in because
the JavaScript used the keyframe name that was in CSS, which would be
fine but the CSS build step minified it so that if the CSS didn't use it
then it was ignored. I tried adding it into the CSS which made it not be
ignored but the minifier changed the name to shorten it which still
caused the JS to not be able to find the correct keyframes.

The fix here puts the 3 keyframes from spin.js v4 into
_PageLoadingIndicator partial.