This is a refactor of the frontend project.
When it was first made I focussed on using what we had gained from
forking NHS.UK which included a lot of best practises such as component
library etc. It turns out that no one ever uses any of this stuff so I
want to minimise the confusion to make it as easy as possible to work
on.

So this commit removes a lot of unused stuff and renames the folder
structure of the rest so it makes more sense at first glance.