Adds timings partial to Other Services page.
In an ideal world the data would come from the database.
Given there are complications with Scala etc for now I've gone the easy
route of adding a switch/case statement with all the primary care dx
codes that use other services page. I'm completely fine with replacing
that as soon as possible.