This changes page order for email and adds bug fixes.
Beforehand you could only click Skip link to have the email address
empty. Now it can be empty but the skip link is still available too.
Apart from nicer UX, this allows the other journeys that do not use the
email page to validate the model without specifically hiding the email
model, so it just validates as empty.