Covid Eliminator
This adds a blue box to dispositions to tell the user they probably
haven't got COVID-19. It only shows on non-covid related dispositions
(ie. didn't end up in the new Health Alert pathway) AND when the user
has been through the Guided Selection page. So if they don't go through
that, this will not show. It will show if they go through it but answer
no.

Note: The _CovidEliminator partial must be put within a container with
`.nhsuk-core` on it, or the styling will go wrong within the box.