This moves the CSS for keyframe animation for spinner
This was originally in the partial as a workaround for the CSS minifier however this change instead disables the minification of the keyframe property, removing the need for the CSS to be in the partial. The reason for doing this is CSS style element inside body is invalid HTML so fails the WCAG parsing criterion.